Move code generator to "codegen" subfolder
