+++ lib/codegen/generator-base.js
