+++ lib/codegen/generator-v1.2.js
