+++ lib/codegen/spec-converter.js
