+++ test/codegen/swagger-v12.test.js
@@ -1,5 +1,5 @@
 var expect = require('chai').expect;
-var V12Generator = require('../lib/generator-v1.2.js');
+var V12Generator = require('../../lib/codegen/generator-v1.2.js');
 
 function generate(spec, options) {
   var generator = new V12Generator();
@@ -9,7 +9,7 @@ function generate(spec, options) {
 describe('Swagger spec v1.2 generator', function () {
 
   it('generates remote methods', function () {
-    var petStoreV12Spec = require('../example/pet-store-1.2.json');
+    var petStoreV12Spec = require('../../example/pet-store-1.2.json');
     var code = generate(petStoreV12Spec, {modelName: 'Store'});
     expect(code).to.be.string;
   });
